[QI2-1098] Add e2e test and allow to pass
