+++ pennylane_quantuminspire2/api/pagination.py
@@ -1,0 +1,3 @@
+from qiskit_quantuminspire.api.pagination import PageReader
+
+__all__ = ['PageReader']
