+++ pennylane_quantuminspire2/api/settings.py
@@ -1,0 +1,3 @@
+from qiskit_quantuminspire.api.settings import ApiSettings, AuthSettings, TokenInfo
+
+__all__ = ['ApiSettings', 'AuthSettings', 'TokenInfo']
