+++ pyproject.toml
@@ -21,6 +21,7 @@ python = "^3.9"
 pennylane-qiskit = "^0.38.1"
 qiskit-quantuminspire = "^0.1.0"
 qi-compute-api-client = "^0.35.0"
+qiskit-quantuminspire = "^0.1.0"
 
 [tool.poetry.group.dev.dependencies]
 requests = "^2.32.3"
